Move async traits out of async-callbacks.
