+++ fixtures/async-traits/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "uniffi-fixtures-async-traits"
+version = "0.21.0"
+authors = ["jhugman"]
+edition = "2021"
+license = "MPL-2.0"
+publish = false
+
+[lib]
+name = "async_traits"
+crate-type = ["lib", "cdylib"]
+
+[dependencies]
+uniffi = { workspace = true, features = ["tokio", "cli"] }
+async-trait = "0.1"
+futures = "0.3"
+thiserror = "1.0"
+tokio = { version = "1.24.1", features = ["time", "sync"] }
+once_cell = "1.18.0"
+ubrn_testing = { path = "../../crates/ubrn_testing" }
+
+[build-dependencies]
+uniffi = { workspace = true, features = ["build"] }
+
+[dev-dependencies]
+uniffi = { workspace = true, features = ["bindgen-tests"] }
