+++ fixtures/async-traits/src/lib.rs
@@ -1,0 +1,51 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/
+ */
+
+use std::{sync::Arc, time::Duration};
+
+use ubrn_testing::timer::{TimerFuture, TimerService};
+
+// Example of an trait with async methods
+#[uniffi::export]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+pub trait SayAfterTrait: Send + Sync {
+    async fn say_after(&self, ms: u16, who: String) -> String;
+}
+
+struct SayAfterImpl1;
+
+struct SayAfterImpl2;
+
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+impl SayAfterTrait for SayAfterImpl1 {
+    async fn say_after(&self, ms: u16, who: String) -> String {
+        say_after(ms, who).await
+    }
+}
+
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
+impl SayAfterTrait for SayAfterImpl2 {
+    async fn say_after(&self, ms: u16, who: String) -> String {
+        say_after(ms, who).await
+    }
+}
+
+#[uniffi::export]
+fn get_say_after_traits() -> Vec<Arc<dyn SayAfterTrait>> {
+    vec![Arc::new(SayAfterImpl1), Arc::new(SayAfterImpl2)]
+}
+
+/// Async function that says something after a certain time.
+async fn say_after(ms: u16, who: String) -> String {
+    TimerFuture::sleep(Duration::from_millis(ms.into())).await;
+
+    format!("Hello, {who}!")
+}
+
+uniffi::setup_scaffolding!();
