+++ fixtures/async-traits/tests/bindings/.supported-flavors.txt
@@ -1,0 +1,1 @@
+jsi
