+++ fixtures/async-traits/uniffi.toml
@@ -1,0 +1,3 @@
+[bindings.typescript]
+logLevel = "debug"
+consoleImport = "@/hermes"
